% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baw_links.R
\name{add_download_link}
\alias{add_download_link}
\title{Add Download Link}
\usage{
add_download_link(df, host, arid, start, duration)
}
\arguments{
\item{df}{A data frame containing information about audio files.}

\item{host}{The hostname (e.g., "api.acousticobservatory.org") where the audio files are hosted.}

\item{arid}{A character or integer vector of audio recording IDs.}

\item{start}{A numeric vector of start offsets (in seconds) for each audio file.}

\item{duration}{A single number or numeric vector of durations (in seconds) for each audio file.}
}
\value{
The modified data frame with a new `download` column containing the download links.
}
\description{
Adds a column with a download link for audio files hosted on a specific server.
}
