% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baw_links.R
\name{add_listen_link}
\alias{add_listen_link}
\title{Add Listen Link}
\usage{
add_listen_link(df, host, arid, start, duration)
}
\arguments{
\item{df}{A data frame containing information about audio files.}

\item{host}{The hostname (e.g., "example.com") where the audio files are hosted.}

\item{arid}{A character vector of acoustic recording IDs.}

\item{start}{A numeric vector of start offsets (in seconds) for each audio file.}

\item{duration}{A numeric vector of durations (in seconds) for each audio file.}
}
\value{
The modified data frame with a new `listen` column containing the listen links.
}
\description{
Adds a column with a link to listen to audio files hosted on a specific server.
}
