% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird_observations}
\alias{bird_observations}
\title{Bird Observation Data}
\format{
## `who`
A data frame with 5000 rows and 6 columns:
\describe{
  \item{site}{site name (factor)}
  \item{arid}{int}
  \item{offset_seconds}{Offset from start of file in seconds}
  \item{timestamp}{datetime of start of detection}
  \item{label}{Label (factor)}
  \item{score}{}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
bird_observations
}
\description{
A subset of bird detections with label and site names changed
}
\keyword{datasets}
