% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchServer}
\alias{launchServer}
\title{Launch Shiny Server}
\usage{
launchServer(data, config = list(), test_subset = NA)
}
\arguments{
\item{data}{A data.frame or tibble, or a string specifying the path to the RDS file,
containing the classification results data.}

\item{config}{A list containing configuration options for the Shiny app (e.g., API and web host URLs, recording duration).}

\item{test_subset}{A numeric value between 0 and 1 specifying the proportion of the data to use for testing purposes.
Default is NA, meaning all data will be used.}
}
\value{
A Shiny application object.
}
\description{
This function launches a Shiny server to serve the Bird Audio Window results viewer.
}
\examples{
\dontrun{
data_path <- system.file("extdata", "minjerribah_birdnet.rds", package = "bawresultsviewer")

config <- list(
  api_host = "api.acousticobservatory.org",
  web_host = "data.acousticobservatory.org"
)

launchServer(data_path, config)
}

}
