% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{formatResults}
\alias{formatResults}
\title{Format Results}
\usage{
formatResults(
  source,
  col_map = list(filename = "filename", offset_seconds = "timestamp_s"),
  random_subset = NA
)
}
\arguments{
\item{source}{The path to the RDS file containing the results data.}

\item{col_map}{A named list specifying the column names in the input data.
Default: `list(filename = "filename", offset_seconds = "timestamp_s")`.}

\item{random_subset}{To test quickly, supply a number to reduce the computation time}
}
\value{
A tibble with the original data and new columns:
  * `site`: The site name extracted from the filename.
  * `arid`: The ARID extracted from the filename.
  * `timestamp`: The calculated timestamp, combining file start time and offset.
}
\description{
Extracts site, ARID (audio_recording_id), and timestamp information from
a filename column in BAW results data. Assumes the filename follows the
canonical format: `[file_start_timestamp]_[site_name]_[arid].[ext]`.
}
\details{
Generally files analysed from a baw server use a filename containing the start datetime, site and arid. This function assumes the filename column
contains filenames in this format.
}
\examples{
\dontrun{
data_path <- system.file("extdata", "example_baw_results.rds", package = "yourpackage")
formatted_data <- formatResults(data_path)
}
}
