% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchServer}
\alias{launchServer}
\title{Launch Shiny Server}
\usage{
launchServer(data_path, config = list(), test_subset = 0.1)
}
\arguments{
\item{data_path}{A string specifying the path to the RDS file containing the classification results data.}

\item{config}{A list containing configuration options for the Shiny app (e.g., API and web host URLs, recording duration).}
}
\value{
A Shiny application object.
}
\description{
This function launches a Shiny server to serve the Bird Audio Window results viewer.
}
\examples{
\dontrun{
data_path <- system.file("extdata", "minjerribah_birdnet.rds", package = "bawresultsviewer")

config <- list(
  api_host = "api.acousticobservatory.org",
  web_host = "data.acousticobservatory.org"
)

launchServer(data_path, config)
}

}
